(PC-30257)[API] refactor: Simplify `update_venue`
`update_venue` is used in two different contexts and try to do a lot at the same time.
It is called `update_venue` but do validation, deal with admin users (or not) logic, update contact,
update opening hours, keep tracks of all of this and has NINE parameters.
In another words `update_venue` smells a lot and it's getting more and more painfull to deal with it.
This is an attempt of simplification, but this far from enough.

- admin_user:
	The function is called only twice, from the BO and from pcapi. On the first
	case we can trust user input, in the later we can't. First simplification
	occurs here, we don't need `admin_update` parameter, the controller
	can simply call `check_venue_edition` where it's needed.

- attrs & modifications:
	`modifications` is the difference between venue instance field and
	content of the form submit by the user (or from the payload in pcapi)
	The only reason we are passing an extra parameter is for
	`new_permanent` value. `modifications` can replace `attrs`
	to determine weither or not the venue is "newly permanent".
