+++ api/tests/core/offerers/test_api.py
@@ -456,302 +456,6 @@ class DeleteVenueTest:
         assert educational_models.AdageVenueAddress.query.count() == 0
 
 
-class EditVenueTest:
-    @patch("pcapi.core.search.async_index_offers_of_venue_ids")
-    def when_changes_on_public_name_algolia_indexing_is_triggered(self, mocked_async_index_offers_of_venue_ids):
-        # Given
-        user = users_factories.UserFactory()
-        venue = offerers_factories.VenueFactory(
-            name="old names",
-            publicName="old name",
-            city="old City",
-        )
-
-        # When
-        json_data = {"publicName": "my new name"}
-        offerers_api.update_venue(venue, author=user, **json_data)
-
-        # Then
-        mocked_async_index_offers_of_venue_ids.assert_called_once_with(
-            [venue.id],
-            reason=search.IndexationReason.VENUE_UPDATE,
-            log_extra={"changes": {"publicName"}},
-        )
-
-    @patch("pcapi.core.search.async_index_offers_of_venue_ids")
-    def when_changes_on_city_algolia_indexing_is_triggered(self, mocked_async_index_offers_of_venue_ids):
-        # Given
-        user = users_factories.UserFactory()
-        venue = offerers_factories.VenueFactory(
-            name="old names",
-            publicName="old name",
-            city="old City",
-        )
-
-        # When
-        json_data = {"city": "My new city"}
-        offerers_api.update_venue(venue, author=user, **json_data)
-
-        # Then
-        mocked_async_index_offers_of_venue_ids.assert_called_once_with(
-            [venue.id],
-            reason=search.IndexationReason.VENUE_UPDATE,
-            log_extra={"changes": {"city"}},
-        )
-
-    @patch("pcapi.core.search.async_index_offers_of_venue_ids")
-    def when_changes_are_not_on_algolia_fields_it_should_not_trigger_indexing(
-        self, mocked_async_index_offers_of_venue_ids
-    ):
-        # Given
-        user = users_factories.UserFactory()
-        venue = offerers_factories.VenueFactory(
-            name="old names",
-            publicName="old name",
-            city="old City",
-            bookingEmail="old@email.com",
-        )
-
-        # When
-        json_data = {"bookingEmail": "new@email.com"}
-        offerers_api.update_venue(venue, author=user, **json_data)
-
-        # Then
-        mocked_async_index_offers_of_venue_ids.assert_not_called()
-
-    @patch("pcapi.core.search.async_index_offers_of_venue_ids")
-    def when_changes_in_payload_are_same_as_previous_it_should_not_trigger_indexing(
-        self, mocked_async_index_offers_of_venue_ids
-    ):
-        # Given
-        user = users_factories.UserFactory()
-        venue = offerers_factories.VenueFactory(
-            name="old names",
-            publicName="old name",
-            city="old City",
-        )
-
-        # When
-        json_data = {"city": "old City"}
-        offerers_api.update_venue(venue, author=user, **json_data)
-
-        # Then
-        mocked_async_index_offers_of_venue_ids.assert_not_called()
-
-    def test_empty_siret_is_editable(self, app) -> None:
-        # Given
-        user = users_factories.UserFactory()
-        venue = offerers_factories.VenueWithoutSiretFactory()
-
-        venue_data = {
-            "siret": venue.managingOfferer.siren + "11111",
-        }
-
-        # when
-        updated_venue = offerers_api.update_venue(venue, author=user, **venue_data)
-
-        # Then
-        assert updated_venue.siret == venue_data["siret"]
-
-    @pytest.mark.parametrize(
-        "venue_data", [{"siret": "12345678954321", "name": "newName"}, {"siret": None, "comment": "test"}]
-    )
-    def test_existing_siret_is_editable(self, venue_data, app) -> None:
-        # Given
-        user = users_factories.UserFactory()
-        venue = offerers_factories.VenueFactory(siret="12345678900001", managingOfferer__siren="123456789")
-
-        # when
-        offerers_api.update_venue(venue, author=user, **venue_data)
-
-        # Then
-        assert venue.siret == venue_data["siret"]
-        if venue_data["siret"] is not None:
-            assert venue.name == venue_data["name"]
-
-    def test_remove_siret(self, app) -> None:
-        user = users_factories.UserFactory()
-        venue = offerers_factories.VenueFactory(
-            siret="12345678900001",
-            managingOfferer__siren="123456789",
-        )
-        with pytest.raises(api_errors.ApiErrors) as error:
-            offerers_api.update_venue(venue, author=user, name=venue.name, siret="")
-        assert error.value.errors["siret"] == ["Vous ne pouvez pas supprimer le siret d'un lieu"]
-
-    @pytest.mark.parametrize("venue_data", [{"name": "New name"}, {"name": None}])
-    def test_name_is_not_editable(self, venue_data, app) -> None:
-        # Given
-        user = users_factories.UserFactory()
-        venue = offerers_factories.VenueFactory()
-
-        # when
-        with pytest.raises(api_errors.ApiErrors) as error:
-            offerers_api.update_venue(venue, author=user, **venue_data)
-
-        # Then
-        assert error.value.errors["name"] == ["Vous ne pouvez pas modifier la raison sociale d'un lieu"]
-
-    def test_accessibility_fields_are_updated(self, app) -> None:
-        # given
-        user = users_factories.UserFactory()
-        venue = offerers_factories.VenueFactory()
-
-        # when
-        venue_data = {
-            "audioDisabilityCompliant": True,
-            "mentalDisabilityCompliant": True,
-            "motorDisabilityCompliant": False,
-            "visualDisabilityCompliant": False,
-        }
-
-        offerers_api.update_venue(venue, author=user, **venue_data)
-
-        venue = offerers_models.Venue.query.get(venue.id)
-        assert venue.audioDisabilityCompliant
-        assert venue.mentalDisabilityCompliant
-        assert venue.motorDisabilityCompliant is False
-        assert venue.visualDisabilityCompliant is False
-
-    def test_no_modifications(self, app) -> None:
-        # given
-        user = users_factories.UserFactory()
-        venue = offerers_factories.VenueFactory()
-
-        # when
-        venue_data = {
-            "departementCode": venue.departementCode,
-            "city": venue.city,
-            "motorDisabilityCompliant": venue.motorDisabilityCompliant,
-        }
-        contact_data = serialize_base.VenueContactModel(
-            email=venue.contact.email,
-            phone_number=venue.contact.phone_number,
-            social_medias=venue.contact.social_medias,
-            website=venue.contact.website,
-        )
-
-        # Check FF Api Adresse
-        with assert_num_queries(1):
-            offerers_api.update_venue(venue, contact_data=contact_data, author=user, **venue_data)
-
-        assert history_models.ActionHistory.query.count() == 0
-
-    @patch("pcapi.connectors.virustotal.request_url_scan")
-    def test_update_only_venue_contact(self, mock_request_url_scan):
-        user_offerer = offerers_factories.UserOffererFactory(
-            user__email="user.pro@test.com",
-        )
-        venue = offerers_factories.VenueFactory(managingOfferer=user_offerer.offerer, contact=None)
-
-        contact_data = serialize_base.VenueContactModel(email="other.contact@venue.com", phone_number="0788888888")
-
-        offerers_api.update_venue(venue, contact_data=contact_data, author=user_offerer.user)
-
-        venue = offerers_models.Venue.query.get(venue.id)
-        assert venue.contact
-        assert venue.contact.phone_number == contact_data.phone_number
-        assert venue.contact.email == contact_data.email
-        mock_request_url_scan.assert_not_called()
-
-    @patch("pcapi.connectors.virustotal.request_url_scan")
-    def test_update_only_venue_website(self, mock_request_url_scan):
-        user_offerer = offerers_factories.UserOffererFactory(
-            user__email="user.pro@test.com",
-        )
-        venue = offerers_factories.VenueFactory(managingOfferer=user_offerer.offerer)
-
-        contact_data = serialize_base.VenueContactModel(
-            email=venue.contact.email, phone_number=venue.contact.phone_number, website="https://new.website.com"
-        )
-
-        offerers_api.update_venue(venue, contact_data=contact_data, author=user_offerer.user)
-
-        venue = offerers_models.Venue.query.get(venue.id)
-        assert venue.contact.phone_number == "+33102030405"
-        assert venue.contact.email == "contact@venue.com"
-        assert venue.contact.website == "https://new.website.com"
-        mock_request_url_scan.assert_called_once_with(contact_data.website, skip_if_recent_scan=True)
-
-    def test_no_venue_contact_created_if_no_data(self, app):
-        user_offerer = offerers_factories.UserOffererFactory(
-            user__email="user.pro@test.com",
-        )
-        venue = offerers_factories.VenueFactory(managingOfferer=user_offerer.offerer, contact=None)
-        empty_contact_data = serialize_base.VenueContactModel()
-
-        offerers_api.update_venue(venue, contact_data=empty_contact_data, author=user_offerer.user)
-
-        assert offerers_models.VenueContact.query.count() == 0
-
-    def test_cannot_update_virtual_venue_name(self):
-        user = users_factories.UserFactory()
-        offerer = offerers_factories.OffererFactory(siren="000000000")
-        virtual_venue = offerers_factories.VirtualVenueFactory(managingOfferer=offerer)
-
-        venue_data = {"name": "Toto"}
-        with pytest.raises(api_errors.ApiErrors) as error:
-            offerers_api.update_venue(virtual_venue, author=user, **venue_data)
-
-        msg = "Vous ne pouvez pas modifier un lieu Offre Numérique."
-        assert error.value.errors == {"venue": [msg]}
-
-    @patch("pcapi.connectors.virustotal.request_url_scan")
-    def test_no_venue_contact_no_modification(self, mock_request_url_scan) -> None:
-        # given
-        user = users_factories.UserFactory()
-        venue = offerers_factories.VenueFactory(contact=None)
-
-        # when
-        venue_data = {
-            "departementCode": venue.departementCode,
-            "city": venue.city,
-        }
-        # same behavior as when update_venue() is called from PC Pro API
-        contact_data = serialize_base.VenueContactModel(email=None, phone_number=None, social_medias=None, website=None)
-
-        offerers_api.update_venue(venue, contact_data=contact_data, author=user, **venue_data)
-
-        # then
-        # nothing has changed => nothing to save nor update
-        assert history_models.ActionHistory.query.count() == 0
-        mock_request_url_scan.assert_not_called()
-
-    @patch("pcapi.connectors.virustotal.request_url_scan")
-    def test_no_venue_contact_add_contact(self, mock_request_url_scan) -> None:
-        # given
-        user = users_factories.UserFactory()
-        venue = offerers_factories.VenueFactory(contact=None)
-
-        # when
-        venue_data = {
-            "departementCode": venue.departementCode,
-            "city": venue.city,
-        }
-        # same behavior as when update_venue() is called from PC Pro API
-        contact_data = serialize_base.VenueContactModel(
-            email="added@venue.com", phone_number=None, social_medias=None, website="https://www.venue.com"
-        )
-
-        offerers_api.update_venue(venue, contact_data=contact_data, author=user, **venue_data)
-
-        # then
-        assert venue.contact
-        assert venue.contact.email == contact_data.email
-        assert venue.contact.website == contact_data.website
-
-        # contact info added => an action should be logged in history
-        action = history_models.ActionHistory.query.one()
-        assert action.actionType == history_models.ActionType.INFO_MODIFIED
-        assert action.authorUserId == user.id
-        assert action.venueId == venue.id
-        assert action.extraData["modified_info"] == {
-            "contact.email": {"new_info": contact_data.email, "old_info": None},
-            "contact.website": {"new_info": contact_data.website, "old_info": None},
-        }
-        mock_request_url_scan.assert_called_once_with(contact_data.website, skip_if_recent_scan=True)
-
-
 class EditVenueContactTest:
     def test_create_venue_contact(self, app):
         user_offerer = offerers_factories.UserOffererFactory(
